Backend/requirements.txt
Django~=5.1.2
psycopg2
djangorestframework
PyJWT
django-cors-headers